requirements.txt
asgiref==3.8.1
chardet==5.2.0
colorama==0.4.6
Django==5.1.4
djangorestframework==3.15.2
pillow==11.1.0
qrcode==8.0
reportlab==4.3.0
sqlparse==0.5.3
tzdata==2024.2
